fix: resolve all Dart compilation errors and upgrade Android build config
- kdbx_vault.dart: update to kdbx 2.4.2 API (read/write instead of
  decode/encode, entries.remove, drop file: param from KdbxGroup.create,
  KdbxKey('Notes') instead of KdbxKeyCommon.NOTES, pass Uint8List directly
  to Credentials.composite instead of wrapping in KeyFileCredentials)
- argon2_kdf.dart: fix import to use pointycastle instead of argon2_ffi_base
- app_theme.dart: CardTheme→CardThemeData, DialogTheme→DialogThemeData (Flutter 3.44)
- biometric_setup_screen.dart: add missing import for masterKeyManagerProvider
- master_key_manager.dart: remove invalid fillRange() call on String
- password_strength_indicator.dart: cast zxcvbn score to int
- android/settings.gradle: AGP 8.7→8.9.1, Kotlin 1.9.22→2.0.21
- android/app/build.gradle: compileSdk/targetSdk 34→36 (required by deps)
- gradle-wrapper.properties: Gradle 8.11.1→8.14.0

+++ lib/features/onboarding/presentation/widgets/password_strength_indicator.dart
@@ -13,7 +13,7 @@ class PasswordStrengthIndicator extends StatelessWidget {
     if (password.isEmpty) return const SizedBox.shrink();
 
     final result = Zxcvbn().evaluate(password);
-    final score = result.score ?? 0;
+    final score = (result.score ?? 0).toInt();
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
